Include TEXT1 in generated bins, avoid boot loop
The Arduino 8266 code recently split .text into .text and .text1 to
enable placing certain functions in IRAM automatically (required for
callbacks in IRQs).

Update the binary creation and size calculation to account for this.

Without this change, infinite boot loops will occur because the main
code will never be uploaded to the chip.

Fixes #143