Add new board Shelly dimmer 2
The shelly dimmer 2 [0] board is tiny LIVE AC powered board used as a dimmer
by Shelly.

It features a 16M (2mb) quad IO flash chip and already has support in
tasmota [1] (and home-assistant).

Dimming functionality is provided by a secondary STM32 microcontroller,
which is driven via UART0, which means that when using the
microcontroller (and pulling it out of reset), UART0 cannot be used for
debugging. Flashing works just fine, as the STM32 is in reset and
ignores the serial messages. Runtime debugging is provided by the
exposed UART1_TX pin (labeled ESP_DBG_TX by shelly).

[0]: https://us.shelly.com/products/shelly-dimmer2
[1]: https://templates.blakadder.com/shelly_Dimmer_2.html

Signed-off-by: Olliver Schinagl <oliver@schinagl.nl>